+++ tensornetwork/utils.py
@@ -13,8 +13,6 @@
 # limitations under the License.
 
 import h5py
-
-from tensornetwork import ops
 from tensornetwork.network import TensorNetwork
 from tensornetwork.component_factory import get_component
 from tensornetwork.network_components import Edge, BaseNode
@@ -148,23 +146,3 @@ def load(path: str):
       edge_data = net_file["edges/" + edge]
       Edge._load_edge(edge_data, nodes_dict)
   return net
-
-
-class NodeCollection:
-
-  def __init__(self, container):
-    if not isinstance(container, (list, set)):
-      raise ValueError("Item passed to NodeCollection must be list or set")
-    self._container = container
-
-  def add(self, node):
-    if isinstance(self._container, set):
-      self._container.add(node)
-    else:
-      self._container.append(node)
-
-  def __enter__(self):
-    ops._default_collection_stack.stack.append(self)
-
-  def __exit__(self, exc_type, exc_val, exc_tb):
-    ops._default_collection_stack.stack.pop()
